requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
crispy-bootstrap5==0.7
cryptography==40.0.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.54.0
django-bootstrap-v5==1.0.11
django-bootstrap5==23.1
django-crispy-forms==2.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-mptt==0.14.0
django-mptt-admin==2.4.1
django-nested-inline==0.4.6
django-nested-inlines==0.1
graphviz==0.20.1
idna==3.4
oauthlib==3.2.2
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.21
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
pyscopg2==66.0.2
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
soupsieve==2.4
sqlparse==0.4.3
urllib3==1.26.15
